Move filter-related modules to filters directory
